Update ABI-32bits and v1.1 uint32_t changes
Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM platform)

+++ core/include/tee/tee_svc_storage.h
@@ -60,7 +60,7 @@ TEE_Result tee_svc_storage_reset_enum(uint32_t obj_enum);
 TEE_Result tee_svc_storage_start_enum(uint32_t obj_enum, uint32_t storage_id);
 
 TEE_Result tee_svc_storage_next_enum(uint32_t obj_enum, TEE_ObjectInfo *info,
-				     void *obj_id, size_t *len);
+				     void *obj_id, uint32_t *len);
 
 /*
  * Data Stream Access Functions
